python/requirements.txt
# Python runtime dependencies (installed into local .venv by Electron PythonService)
Pillow>=10.0.0,<12.0.0
